+++ NEWS.md
@@ -4,6 +4,8 @@
 
  * Add `handler_rpushbullet()` for reporting on progress via the
    Pushbullet Messaging Service using the **RPushbullet** package.
+   
+ * The **notifier** handler reports on interrupts.
 
 ## Bug Fixes